total redesign - whole p5 sketch in one component
